refactor(dev): split up browser and server builds into their own modules
