+++ packages/remix-dev/compiler/compile-browser.ts
@@ -1,0 +1,94 @@
+import * as path from "path";
+import { builtinModules as nodeBuiltins } from "module";
+import * as esbuild from "esbuild";
+import { NodeModulesPolyfillPlugin } from "@esbuild-plugins/node-modules-polyfill";
+import { pnpPlugin as yarnPnpPlugin } from "@yarnpkg/esbuild-plugin-pnp";
+
+import { type BuildOptions } from "../build";
+import { type RemixConfig } from "../config";
+import { getAppDependencies } from "./dependencies";
+import { loaders } from "./loaders";
+import { browserRouteModulesPlugin } from "./plugins/browserRouteModulesPlugin";
+import { cssFilePlugin } from "./plugins/cssFilePlugin";
+import { emptyModulesPlugin } from "./plugins/emptyModulesPlugin";
+import { mdxPlugin } from "./plugins/mdx";
+import { urlImportsPlugin } from "./plugins/urlImportsPlugin";
+
+///////////////////////////////////////////////////////////////////////////////
+
+export async function createBrowserBuild(
+  config: RemixConfig,
+  options: BuildOptions & { incremental?: boolean }
+): Promise<esbuild.BuildResult> {
+  // For the browser build, exclude node built-ins that don't have a
+  // browser-safe alternative installed in node_modules. Nothing should
+  // *actually* be external in the browser build (we want to bundle all deps) so
+  // this is really just making sure we don't accidentally have any dependencies
+  // on node built-ins in browser bundles.
+  let dependencies = Object.keys(getAppDependencies(config));
+  let externals = nodeBuiltins.filter((mod) => !dependencies.includes(mod));
+  let fakeBuiltins = nodeBuiltins.filter((mod) => dependencies.includes(mod));
+
+  if (fakeBuiltins.length > 0) {
+    throw new Error(
+      `It appears you're using a module that is built in to node, but you installed it as a dependency which could cause problems. Please remove ${fakeBuiltins.join(
+        ", "
+      )} before continuing.`
+    );
+  }
+
+  let entryPoints: esbuild.BuildOptions["entryPoints"] = {
+    "entry.client": path.resolve(config.appDirectory, config.entryClientFile),
+  };
+  for (let id of Object.keys(config.routes)) {
+    // All route entry points are virtual modules that will be loaded by the
+    // browserEntryPointsPlugin. This allows us to tree-shake server-only code
+    // that we don't want to run in the browser (i.e. action & loader).
+    entryPoints[id] = config.routes[id].file + "?browser";
+  }
+
+  let plugins = [
+    cssFilePlugin(options),
+    urlImportsPlugin(),
+    mdxPlugin(config),
+    browserRouteModulesPlugin(config, /\?browser$/),
+    emptyModulesPlugin(config, /\.server(\.[jt]sx?)?$/),
+    NodeModulesPolyfillPlugin(),
+    yarnPnpPlugin(),
+  ];
+
+  return esbuild.build({
+    entryPoints,
+    outdir: config.assetsBuildDirectory,
+    platform: "browser",
+    format: "esm",
+    external: externals,
+    loader: loaders,
+    bundle: true,
+    logLevel: "silent",
+    splitting: true,
+    sourcemap: options.sourcemap,
+    metafile: true,
+    incremental: options.incremental,
+    // As pointed out by https://github.com/evanw/esbuild/issues/2440, when tsconfig is set to
+    // `undefined`, esbuild will keep looking for a tsconfig.json recursively up. This unwanted
+    // behavior can only be avoided by creating an empty tsconfig file in the root directory.
+    tsconfig: config.tsconfigPath,
+    mainFields: ["browser", "module", "main"],
+    treeShaking: true,
+    minify: options.mode === "production",
+    entryNames: "[dir]/[name]-[hash]",
+    chunkNames: "_shared/[name]-[hash]",
+    assetNames: "_assets/[name]-[hash]",
+    publicPath: config.publicPath,
+    define: {
+      "process.env.NODE_ENV": JSON.stringify(options.mode),
+      "process.env.REMIX_DEV_SERVER_WS_PORT": JSON.stringify(
+        config.devServerPort
+      ),
+    },
+    jsx: "automatic",
+    jsxDev: options.mode !== "production",
+    plugins,
+  });
+}
